feat: agent integration overhaul - unified output pipeline
Major refactoring of agent integration architecture:

## Phase 1: Foundation - Contracts and Types
- Add NormalizedEntry types in events/normalized.rs
- Add structured ExecutionError types in errors/execution.rs
- Add new channel constants for raw/normalized/error streams

## Phase 2: Core Pipeline Components
- Create LineBuffer service for byte accumulation and ANSI stripping
- Create EventNormalizer service for UnifiedAgentEvent transformation
- Create PermissionDetector service with multi-pattern detection

## Phase 3: Unified Output Pipeline
- Replace AgentOutputSink with AgentOutputPipeline in orchestrator
- Delete legacy ProcessBuffer (dual buffering eliminated)
- Delete legacy ProcessManager (functionality consolidated)

## Phase 4: Enhanced Permission Handling
- Add comprehensive regex patterns for all prompt formats
- Add permission timeout tracking with auto-timeout mechanism
- Add PermissionStatus::TimedOut variant

## Phase 5: Tool State Consolidation
- Remove in-memory tool tracking (DB is single source of truth)
- Add execution context fields (command, file_path, duration_ms)
- Enhanced tool state service methods

## Phase 6: Streaming Architecture
- Dual WebSocket channels (raw-output, normalized)
- Tool state channel for real-time updates
- Incremental event fetching with afterSequence parameter

## Phase 7: Provider Abstraction Improvements
- Extended AgentProvider trait with new methods
- Updated all providers (claude_code, gemini_cli, codex_cli, mock)

## Phase 8: Database Schema Updates
- Migration 016: Enhanced permissions table
- Migration 017: Permission timed out status
- Migration 018: Enhanced tool states table
- Migration 019: Normalized events table

## Phase 9: Frontend Updates
- New TypeScript types generated
- New hooks for dual streams (useAgentRawStream, useAgentNormalizedEvents)
- Updated event subscriptions for new channels

## Phase 10: Cleanup and Validation
- Removed all legacy code (process_buffer.rs, process_manager.rs)
- Updated module exports
- Updated CLAUDE.md documentation
- Added integration tests

+++ crates/openflow-server/src/routes/agents.rs
@@ -132,7 +132,7 @@ async fn list_running(State(state): State<AppState>) -> ServerResult<Json<Vec<Ag
 #[serde(rename_all = "camelCase")]
 pub struct GetEventsQuery {
     /// Only return events with sequence > this value
-    pub after_sequence: Option<i64>,
+    pub after_sequence: Option<i32>,
 }
 
 /// GET /api/agents/sessions/{id}/events
@@ -153,7 +153,7 @@ async fn get_session_events(
 async fn get_latest_sequence(
     State(state): State<AppState>,
     Path(id): Path<String>,
-) -> ServerResult<Json<Option<i64>>> {
+) -> ServerResult<Json<Option<i32>>> {
     let sequence = agent_session::get_latest_sequence(&state.pool, &id).await?;
     Ok(Json(sequence))
 }
@@ -746,7 +746,7 @@ mod tests {
         let body = axum::body::to_bytes(response.into_body(), usize::MAX)
             .await
             .unwrap();
-        let sequence: Option<i64> = serde_json::from_slice(&body).unwrap();
+        let sequence: Option<i32> = serde_json::from_slice(&body).unwrap();
         assert!(sequence.is_none());
 
         // Add events
@@ -773,7 +773,7 @@ mod tests {
         let body = axum::body::to_bytes(response.into_body(), usize::MAX)
             .await
             .unwrap();
-        let sequence: Option<i64> = serde_json::from_slice(&body).unwrap();
+        let sequence: Option<i32> = serde_json::from_slice(&body).unwrap();
         assert_eq!(sequence, Some(1));
     }
